+++ src/components/views/charts/VerticalTable.js
@@ -8,17 +8,17 @@ import {
   CardHeader,
   CardBody
 } from 'shards-react'
-import useResponsiveStyles from '../../hooks/useResponsiveStyles'
-import startOverSvg from '../../assets/svgs-startover.svg'
+import useResponsiveStyles from '../../../hooks/useResponsiveStyles'
+import startOverSvg from '../../../assets/svgs-startover.svg'
 
 
-export default function HorizontalTable(props) {
-  const { greeting, headers, firstRoundAtt, firstRoundT, lastRoundAtt, lastRoundT, bestRoundAtt, bestRoundT, startOver } = props
+export default function VerticalTable(props) {
+  const { greeting, verticalTableAtt, verticalTableT, startOver } = props
   const sizedStyles = useResponsiveStyles()
   const { borderRadius, sessionTitle } = sizedStyles
 
   return (
-    <Container fluid className="main-content-container px-4" id='container'style={{backgroundColor: 'black', minHeight: '150vh', fontFamily: "'Overpass Mono', monospace"}}>
+    <Container fluid className="main-content-container px-4" id='container'style={{backgroundColor: 'black', minHeight: '100vh', fontFamily: "'Overpass Mono', monospace"}}>
       <Row style={{display: 'flex', justifyContent: 'center'}} noGutters>
         <Col sm='12' lg='8' style={{border: '5px solid black', borderRadius: borderRadius, marginLeft: '5%', marginRight: '5%', marginTop: '2%', backgroundColor: '#e5e6eb', fontFamily: "'Overpass Mono', monospace"}}>
           <Row style={{display: 'flex', justifyContent: 'center', margin: '5%'}}><h2 style={sessionTitle}>Session Status</h2></Row>
@@ -31,7 +31,7 @@ export default function HorizontalTable(props) {
                 <Col>
                   <Card small className="mb-4">
                     <CardHeader className="border-bottom">
-                      <h5 className="m-0" style={{textAlign: 'center', fontFamily: "'Overpass Mono', monospace", fontWeight: '600'}}>ATTEMPTS</h5>
+                      <h5 className="m-0" style={{textAlign: 'center', fontFamily: "'Overpass Mono', monospace", fontWeight: '800'}}>ATTEMPTS</h5>
                     </CardHeader>
                     <CardBody className="p-0 pb-3">
                       <table className="table mb-0">
@@ -39,22 +39,19 @@ export default function HorizontalTable(props) {
                           <tr>
                             <th scope="col" className="border-0">
                             </th>
-                            {headers}
+                            <th scope="col" className="border-0">
+                              1ST
+                            </th>
+                            <th scope="col" className="border-0">
+                              LAST
+                            </th>
+                            <th scope="col" className="border-0">
+                              BEST
+                            </th>
                           </tr>
                         </thead>
                         <tbody>
-                          <tr>
-                            <td>1ST</td>
-                            {firstRoundAtt}
-                          </tr>
-                          <tr>
-                            <td>LAST</td>
-                            {lastRoundAtt}
-                          </tr>
-                          <tr>
-                            <td>BEST</td>
-                            {bestRoundAtt}
-                          </tr>
+                          {verticalTableAtt}
                         </tbody>
                       </table>
                     </CardBody>
@@ -65,30 +62,27 @@ export default function HorizontalTable(props) {
                 <Col>
                   <Card small className="mb-4">
                     <CardHeader className="border-bottom">
-                      <h5 className="m-0" style={{textAlign: 'center', fontFamily: "'Overpass Mono', monospace", fontWeight: '600'}}>TIMES</h5>
+                      <h5 className="m-0" style={{textAlign: 'center', fontFamily: "'Overpass Mono', monospace", fontWeight: '800'}}>TIMES</h5>
                     </CardHeader>
                     <CardBody className="p-0 pb-3">
                       <table className="table mb-0">
                         <thead className="bg-light">
-                          <tr>
-                            <th scope="col" className="border-0">
-                            </th>
-                            {headers}
-                          </tr>
-                        </thead>
-                        <tbody>
-                          <tr>
-                            <td>1ST</td>
-                            {firstRoundT}
-                          </tr>
-                          <tr>
-                            <td>LAST</td>
-                            {lastRoundT}
-                          </tr>
-                          <tr>
-                            <td>BEST</td>
-                            {bestRoundT}
-                          </tr>
+                        <tr>
+                          <th scope="col" className="border-0">
+                          </th>
+                          <th scope="col" className="border-0">
+                            1ST
+                          </th>
+                          <th scope="col" className="border-0">
+                            LAST
+                          </th>
+                          <th scope="col" className="border-0">
+                            BEST
+                          </th>
+                        </tr>
+                      </thead>
+                      <tbody>
+                        {verticalTableT}
                         </tbody>
                       </table>
                     </CardBody>
@@ -98,9 +92,9 @@ export default function HorizontalTable(props) {
             </Col>
           </Row>
             <Row style={{display: 'flex', justifyContent: 'center', margin: '5%'}}>
-                <Button theme="dark" style={{display: 'block', cursor: 'pointer', backgroundColor: '#e5e6eb', border: 'none', padding: '0'}} onClick={(e) => {
-                  startOver(true)
-                }}><img src={startOverSvg} alt='next round' style={{width: '15rem'}}></img></Button>
+              <Button theme="dark" style={{display: 'block', cursor: 'pointer', backgroundColor: '#e5e6eb', border: 'none', padding: '0'}} onClick={(e) => {
+                startOver(true)
+              }}><img src={startOverSvg} alt='next round' style={{width: '15rem'}}></img></Button>
             </Row>
           </Col>
         </Row>
